Optimize schemas from routes
Centralize and optimize schemas for API responses.

* Add `src/schemas/bca.ts` to define schemas for BCA exchange rate details, data structure, success responses, and error responses.
* Add `src/schemas/bi.ts` to define schemas for BI exchange rate details, data structure, success responses, and error responses.
* Add `src/schemas/mandiri.ts` to define schemas for Mandiri exchange rate details, data structure, success responses, and error responses.
* Modify `src/routes/bca.ts` to import schemas from `src/schemas/bca.ts` and remove local schema definitions.
* Modify `src/routes/bi.ts` to import schemas from `src/schemas/bi.ts` and remove local schema definitions.
* Modify `src/routes/mandiri.ts` to import schemas from `src/schemas/mandiri.ts` and remove local schema definitions.
* Modify `src/index.ts` to add `hide: true` for path `/` in the `redirect` route.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/k-ardliyan/idr-exchange-rates?shareId=XXXX-XXXX-XXXX-XXXX).
